API

WindowAbstractions.BUTTON_PRESSED

WindowAbstractions.BUTTON_RELEASED

WindowAbstractions.KEY_PRESSED

WindowAbstractions.KEY_RELEASED

WindowAbstractions.POINTER_ENTERED

WindowAbstractions.POINTER_EXITED

WindowAbstractions.POINTER_MOVED

WindowAbstractions.WINDOW_CLOSED

WindowAbstractions.WINDOW_EXPOSED

WindowAbstractions.WINDOW_GAINED_FOCUS

WindowAbstractions.WINDOW_INVALID

WindowAbstractions.WINDOW_LOST_FOCUS

WindowAbstractions.WINDOW_RESIZED

Core.Type

WindowAbstractions.AbstractWindow

WindowAbstractions.AbstractWindowManager

WindowAbstractions.Event

WindowAbstractions.KeyCombination

WindowAbstractions.KeyEvent

WindowAbstractions.KeySymbol

Base.close

WindowAbstractions.dpi

WindowAbstractions.extent

WindowAbstractions.get_window

WindowAbstractions.key_symbol

WindowAbstractions.map_window

WindowAbstractions.poll_for_events!

WindowAbstractions.resize

WindowAbstractions.set_icon

WindowAbstractions.set_icon_title

WindowAbstractions.set_title

WindowAbstractions.unmap_window

WindowAbstractions.windows

WindowAbstractions.@key_str

WindowAbstractions.BUTTON_PRESSED — Constant

A mouse button was released.

WindowAbstractions.BUTTON_RELEASED — Constant

The pointer entered the window area.

WindowAbstractions.KEY_PRESSED — Constant

A key was pressed. Note that some key combinations can be reserved by the OS, so they don't trigger the corresponding event. On Ubuntu 20.04, this is for example the case with some combinations of the form alt+fkey such as alt+f4.

WindowAbstractions.KEY_RELEASED — Constant

A key was released.

WindowAbstractions.POINTER_ENTERED — Constant

A mouse button was pressed.

WindowAbstractions.POINTER_EXITED — Constant

The pointer left the window area.

WindowAbstractions.POINTER_MOVED — Constant

The pointer moves in the window.

WindowAbstractions.WINDOW_CLOSED — Constant

The window is closing normally upon raising an exception of type CloseWindow.

WindowAbstractions.WINDOW_EXPOSED — Constant

A window was exposed to the screen; either right after creation, or when it was previously hidden and then visible again.

WindowAbstractions.WINDOW_GAINED_FOCUS — Constant

The window gained focus.

WindowAbstractions.WINDOW_INVALID — Constant

The window is detected as invalid (for example, when closed externally).

The conditions for tagging a window as invalid depend on the windowing API used. It is typically used to signal that a window crashed or does not exist, but for which an event was reported. It may however not possible to know exactly why or when a window becomes invalid. For example, the X11 protocol does not offer a way to check for window validity, since the window may get invalidated by the time the X server answer comes back. Windows that are tagged invalid should be terminated.

WindowAbstractions.WINDOW_LOST_FOCUS — Constant

The window lost focus.

WindowAbstractions.WINDOW_RESIZED — Constant

The window was resized.

Core.Type — Method

(::Type{<:AbstractWindow})(wm::AbstractWindowManager; kwargs...)

Create a window.

WindowAbstractions.AbstractWindow — Type

Abstract window type.

abstract type AbstractWindow

WindowAbstractions.AbstractWindowManager — Type

Entity that manages windows.

abstract type AbstractWindowManager

WindowAbstractions.Event — Type

Generic event structure identified by its type (see EventType) and which may hold data depending on the type of event.

The type of data follows the association (with respect to the event type):

KEY_PRESSED or KEY_RELEASED: KeyEvent

BUTTON_PRESSED or BUTTON_RELEASED: MouseEvent

POINTER_MOVED: PointerState

WINDOW_RESIZED: Tuple{Float64,Float64} # new dimensions

WINDOW_EXPOSED: Tuple{Float64,Float64} # area to redraw

Other event types: Nothing

WindowAbstractions.KeyCombination — Type

Key binding associated to a character and a key modifier state.

struct KeyCombination

key::KeySymbol

exact_modifiers::ModifierState

significant_modifiers::ModifierState

WindowAbstractions.KeyEvent — Type

Details surrounding an event produced by a keystroke.

struct KeyEvent

key_name::Symbol

The name of the physical key.

key::KeySymbol

The symbol associated with the keystroke.

input::Char

Printable input (can be the empty string).

modifiers::ModifierState

Modifier state.

consumed_modifiers::ModifierState

Modifiers that were used when translating the physical key into a key symbol.

WindowAbstractions.KeySymbol — Type

Representation of a symbol that may be bound to a key, depending on a keyboard layout. It differs from the physical representation of a key, denoted by names such as AD01 which represent physical keys (in this case, q on a US keyboard). A key symbol is sent when a physical key is pressed, even if no input character is bound to the key. This structure provides a friendly symbol field, and a prettier repr field (mostly used for printing). For example, the right arrow, with symbol :right_arrow (resp. :kp_right_arrow for keypad input), is represented by "→" (resp. "→ (keypad)").

There is no consistent representation, so the one provided here may differ from windowing API implementations.

struct KeySymbol

symbol::Symbol

repr::String

Base.close — Method

close(wm::AbstractWindowManager, win::AbstractWindow)

Close a window.

close(wm::AbstractWindowManager, _::AbstractWindow)

WindowAbstractions.dpi — Method

Get window DPI.

dpi(win::AbstractWindow)

WindowAbstractions.extent — Method

Window dimensions (width, height).

extent(win::AbstractWindow)

WindowAbstractions.get_window — Method

Retrieve a window given an identifier id, would typically be an integer identifier, a handle or a name.

get_window(wm::AbstractWindowManager, id)

WindowAbstractions.key_symbol — Method

Define an internal representation and a readable description for some key inputs.

The keys that are listed here are not exhaustive and only cover a very small subset of all possible keys.

key_symbol(name::Symbol) -> KeySymbol

WindowAbstractions.map_window — Method

Map a window to the screen.

map_window(win::AbstractWindow)

WindowAbstractions.poll_for_events! — Method

Non-blocking poll for events. Must return whether a client event was consumed, even if no Event is added to the queue.

poll_for_events!(queue::EventQueue)

WindowAbstractions.resize — Method

Resize the window with extent reflecting the new (width, height).

resize(win::AbstractWindow, extent)

WindowAbstractions.set_icon — Method

Set window icon.

set_icon(win::AbstractWindow, icon)

WindowAbstractions.set_icon_title — Method

Set window icon title.

set_icon_title(win::AbstractWindow, icon_title)

WindowAbstractions.set_title — Method

Set window title.

set_title(win::AbstractWindow, title)

WindowAbstractions.unmap_window — Method

Unmap a window from the screen.

unmap_window(win::AbstractWindow)

WindowAbstractions.windows — Method

Retrieve all the AbstractWindows in use by wm.

windows(wm::AbstractWindowManager)

WindowAbstractions.@key_str — Macro

Construct a KeyCombination instance from a string as "[<state_1>+[...<state_n>+]]<key>". The string must be a list of elements separated by '+' characters, with only one key symbol at the end. For example, "k", "alt+k" and "ctrl+alt+shift+f11" are valid strings, but "k+a" is not. Casing is significant, as z and Z are different key symbols. For modifier strings, the casing should be lowercase, or will be forcefully converted into a lowercase string (reducing performance).