API

Core.Type

WindowAbstractions.AbstractWindow

WindowAbstractions.AbstractWindowManager

WindowAbstractions.Callbacks

WindowAbstractions.CloseWindow

WindowAbstractions.EventDetails

WindowAbstractions.InvalidWindow

WindowAbstractions.KeyCombination

WindowAbstractions.KeyContext

WindowAbstractions.KeyEvent

WindowAbstractions.KeyModifierState

WindowAbstractions.KeySymbol

WindowAbstractions.MouseState

WindowAbstractions.WindowCallbacks

WindowAbstractions.WindowException

Base.close

Base.run

WindowAbstractions.dpi

WindowAbstractions.extent

WindowAbstractions.key_symbol

WindowAbstractions.listen_for_events

WindowAbstractions.map_window

WindowAbstractions.poll_for_event

WindowAbstractions.pressed_buttons

WindowAbstractions.process_event

WindowAbstractions.set_extent

WindowAbstractions.set_icon

WindowAbstractions.set_icon_title

WindowAbstractions.set_title

WindowAbstractions.terminate_window!

WindowAbstractions.unmap_window

WindowAbstractions.wait_for_event

WindowAbstractions.@key_str

Core.Type — Method

(::Type{<:AbstractWindow})(wm::AbstractWindowManager; kwargs...)

Create a window.

WindowAbstractions.AbstractWindow — Type

Abstract window type.

abstract type AbstractWindow

WindowAbstractions.AbstractWindowManager — Type

Entity that manages windows.

abstract type AbstractWindowManager

WindowAbstractions.Callbacks — Type

Abstract type for callback structures. Callback structures should possess function fields with names coherent with the corresponding callback type.

abstract type Callbacks

WindowAbstractions.CloseWindow — Type

Signals that the window must be closed.

struct CloseWindow <: WindowException

win::AbstractWindow

msg::String

WindowAbstractions.EventDetails — Type

Generic event structure holding data as an EventData member.

WindowAbstractions.InvalidWindow — Type

Exception to be raised when a window is deemed invalid. The conditions for tagging a window as invalid depend on the windowing API used. It is typically used to signal that a window crashed or does not exist, but for which an event was reported. It may however not possible to know exactly why or when a window becomes invalid. For example, the X11 protocol does not offer a way to check for window validity, since the window may get invalidated by the time the X server answer comes back. Windows that are tagged invalid should be terminated.

struct InvalidWindow <: WindowException

win::AbstractWindow

msg::String

WindowAbstractions.KeyCombination — Type

Key binding associated to a character and a key modifier state.

struct KeyCombination

key::String

state::KeyModifierState

WindowAbstractions.KeyContext — Type

State of the keyboard, with regards to numlock and caps states.

struct KeyContext

numlock::Bool

caps::Bool

WindowAbstractions.KeyEvent — Type

Event data regarding a keystroke.

WindowAbstractions.KeyModifierState — Type

Provide information about key modifier state (shift, control, alt and the OS key).

struct KeyModifierState

shift::Bool

ctrl::Bool

alt::Bool

super::Bool

WindowAbstractions.KeySymbol — Type

Representation of a symbol that may be bound to a key, depending on a keyboard layout. It differs from the physical representation of a key, denoted by names such as AD01 which represent physical keys (in this case, q on a US keyboard). A key symbol is sent when a physical key is pressed, even if no input character is bound to the key. This structure provides a friendly symbol field, and a prettier repr field (mostly used for printing). For example, the right arrow, with symbol :right_arrow (resp. :kp_right_arrow for keypad input), is represented by "→" (resp. "→ (keypad)").

There is no consistent representation, so the one provided here may differ from windowing API implementations.

struct KeySymbol

symbol::Symbol

repr::String

WindowAbstractions.MouseState — Type

Store mouse input events, including clicking buttons and scrolling.

struct MouseState

left::Bool

middle::Bool

right::Bool

scroll_up::Bool

scroll_down::Bool

any::Bool

WindowAbstractions.WindowCallbacks — Type

Set of common window callbacks, in response to specific window events.

The callbacks on_close and on_invalid take respectively a CloseWindow and an InvalidWindow as argument, while other callbacks take an EventDetails as argument parametrized according to the event.

Each field documents when and in which context the callback may be called.

struct WindowCallbacks <: Callbacks

on_close::Any

The window is closing normally upon raising an exception of type CloseWindow.

on_invalid::Any

The window is detected as invalid (for example, when closed externally) through an exception of type InvalidWindow.

on_resize::Any

The window was resized.

on_mouse_button_pressed::Any

A mouse button was pressed.

on_mouse_button_released::Any

A mouse button was released.

on_pointer_move::Any

The pointer moves in the window.

on_pointer_enter::Any

The pointer enters the window area.

on_pointer_leave::Any

The pointer leaves the window area.

on_key_pressed::Any

A key was pressed. Note that some key combinations can be reserved by the OS, so they don't trigger the corresponding event. On Ubuntu 20.04, this is for example the case with some combinations of the form alt+fkey such as alt+f4.

on_key_released::Any

A key was released.

on_expose::Any

A window was exposed to the screen; either right after creation, or when it was previously hidden and then visible again.

WindowAbstractions.WindowException — Type

Exception attached to a specific AbstractWindow instance.

abstract type WindowException <: Exception

Base.close — Method

close(wm::AbstractWindowManager, win::AbstractWindow)

Close a window.

close(wm::AbstractWindowManager, _::AbstractWindow)

Base.run — Method

Run an event loop for the window manager wm which listens for events and processes them.

If sleep_time is non-zero, Base.sleep will be called between each round of polling (i.e. between successive tries that return no event).

run(wm::AbstractWindowManager; sleep_time)

WindowAbstractions.dpi — Method

Get window DPI.

dpi(win::AbstractWindow)

WindowAbstractions.extent — Method

Window dimensions (width, height).

extent(win::AbstractWindow)

WindowAbstractions.key_symbol — Method

Define an internal representation and a readable description for some key inputs.

The keys that are listed here are not exhaustive and only cover a very small subset of all possible keys.

key_symbol(name::Symbol) -> KeySymbol

WindowAbstractions.listen_for_events — Method

Run an event loop for the window manager wm which listens for events and processes them.

If sleep_time is non-zero, Base.sleep will be called between each round of polling (i.e. between successive tries that return no event).

listen_for_events(wm::AbstractWindowManager; sleep_time)

WindowAbstractions.map_window — Method

Map a window to the screen.

map_window(win::AbstractWindow)

WindowAbstractions.poll_for_event — Method

Non-blocking poll for events. Should return an element of type Nothing if no events were caught.

poll_for_event(wm::AbstractWindowManager)

WindowAbstractions.pressed_buttons — Method

Retrieve a list of pressed button symbols for the left, middle and right buttons from a MouseState.

pressed_buttons(input::MouseState) -> Set{Symbol}

WindowAbstractions.process_event — Method

Process an event received by either poll_for_event or wait_for_event, modifying state (e.g. window, keymap) and/or returning an EventDetails instance.

If the returned value is not an EventDetails, then it must either be of type CloseWindow or InvalidWindow, or nothing.

process_event(wm::AbstractWindowManager, event)

WindowAbstractions.set_extent — Method

Resize the window with extent reflecting the new (width, height).

set_extent(win::AbstractWindow, extent)

WindowAbstractions.set_icon — Method

Set window icon.

set_icon(win::AbstractWindow, icon)

WindowAbstractions.set_icon_title — Method

Set window icon title.

set_icon_title(win::AbstractWindow, icon_title)

WindowAbstractions.set_title — Method

Set window title.

set_title(win::AbstractWindow, title)

WindowAbstractions.terminate_window! — Method

Clean up window state upon destruction. The window cannot be assumed to be mapped or to be valid.

terminate_window!(wm::AbstractWindowManager, win::AbstractWindow)

WindowAbstractions.unmap_window — Method

Unmap a window from the screen.

unmap_window(win::AbstractWindow)

WindowAbstractions.wait_for_event — Method

Wait until the next event occurs.

wait_for_event(wm::AbstractWindowManager)

WindowAbstractions.@key_str — Macro

Construct a KeyCombination instance from a string as "[<state_1>+[...<state_n>+]]<key>". The string must be a list of elements separated by '+' characters, with only one non-state character or one fkey as string. For example, "k", "alt+k" and "ctrl+alt+shift+k" are valid strings, but "k+a" is not. Casing is ignored. Fkeys are simply f<1-12> keys, for example key"alt+f4" is valid.