WindowAbstractions.jl

WindowAbstractions.CloseWindow

WindowAbstractions.dpi

WindowAbstractions.extent

WindowAbstractions.map_window

WindowAbstractions.on_close

WindowAbstractions.on_invalid

WindowAbstractions.set_extent

WindowAbstractions.set_icon

WindowAbstractions.set_icon_title

WindowAbstractions.set_title

WindowAbstractions.terminate

WindowAbstractions.unmap_window

WindowAbstractions.CloseWindow — Type

Signals that the window must be closed.

WindowAbstractions.dpi — Method

Get window DPI

WindowAbstractions.extent — Method

Window dimensions (width, height).

WindowAbstractions.map_window — Method

Map a window to the screen.

WindowAbstractions.on_close — Method

Action to perform when an exception of type CloseWindow is raised.

WindowAbstractions.on_invalid — Method

Action to perform when an exception of type InvalidWindow is raised.

WindowAbstractions.set_extent — Method

Resize the window with extent reflecting the new (width, height).

WindowAbstractions.set_icon — Method

Set window icon.

WindowAbstractions.set_icon_title — Method

Set window icon title.

WindowAbstractions.set_title — Method

Set window title.

WindowAbstractions.terminate — Method

Clean up window state upon destruction. The window cannot be assumed to be mapped or to be valid.

WindowAbstractions.unmap_window — Method

Unmap a window from the screen.